deana_desilva_project1_final.md
*collage image:*
---
https://i.imgur.com/8xj6E4k.jpg

*augmented reality overlay:*
---
https://media.giphy.com/media/YjucRG0lFa0o0U5arC/giphy.gif

*screen recording:*
---
in progress

*documentation:*
---
*title of work:* the squirrel initiative

*dimensions:* 550 x 778 pixels or 7.639 x 10.806 inches

*attribution:*
---
https://i.pinimg.com/originals/41/d8/ee/41d8ee8ccb140a70abbb3035f7721f70.jpg
https://cdn.discordapp.com/attachments/441094630633635840/547285372590358554/download.jpg
https://www.pinterest.es/pin/413697915754032922/

photomosh for glitch effect/ gif rendering 

*Project Description:*
---
the squirrel initiative looks like, at first, a collage of just squirrels. the glitched out version on the squirrel initiative is very distorted and is black and white. although no sound is present, the squirrel initiative pre glitch could sound like any regular day outside. whereas the glitched version of the squirrel initiative may sound like when a tv projects static. along with feeling, the background of the pre glitch squirrel initiative may feel like touching the leaves of a plant, as similar to the ones in the background. versus the glitched version of the squirrel initiative where it may feel like touching a warm tv thats attracted a lot of static on its screen. 
there is no real conceptual or motivational aspect to the squirrel initiative. i simply asked my boyfriend what subject my collage should be on and he chose squirrels.

my influences for this piece would definitely be the squirrels i see in real life roaming around san jose state's campus. the intended context for pre glitch vs glitched squirrel initiative is that the sjsu campus squirrels are simply on a different level in comparison to other squirrels. if you are an sjsu student, professor, or if you just walk through sjsu campus and encounter these squirrels, you may see what i'm trying to depict through the squirrel initiative. one recent influence for this heavy description is that tweet where a squirrel seemed it was trying to ride the electric bikes parked on sjsu campus. the sjsu squirrels are just a lot beefier, they seem as if they are "on something" from the way they present themselves and the way they look in comparison to other, more smaller squirrels. this piece shows a huge 180 in terms of how these squirrels are depicted pre glitch versus glitched.